requirements.txt
# /requirements.txt

# Core dependencies
fastapi
uvicorn
requests
python-dotenv

# Database dependencies
sqlalchemy
psycopg2-binary
alembic

# Scraping dependencies
beautifulsoup4
selenium
lxml

# For Zip Code API
pandas
geopandas
shapely
folium
scikit-learn
numpy
dnspython

# For Machine Learning APIs
openai

# Testing dependencies
pytest
pytest-cov

# For Google APIs
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client